+++ src/init.cc
@@ -1,0 +1,18 @@
+#define R_NO_REMAP
+#include <R.h>
+#include <Rinternals.h>
+#include <R_ext/Rdynload.h>
+#include <R_ext/Visibility.h>
+
+extern "C" SEXP C_toolchain_report(void);
+
+static R_CallMethodDef const CallEntries[] = {
+    {"C_toolchain_report", (DL_FUNC) &C_toolchain_report, 0},
+    {NULL, NULL, 0}
+};
+
+extern "C" attribute_visible void R_init_zusmt(DllInfo * dll) {
+    R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
+    R_useDynamicSymbols(dll, FALSE);
+    R_forceSymbols(dll, TRUE);
+}